broken items

+++ Employee of the month/Assets/Laser/Scripts/Laser.cs
@@ -125,7 +125,7 @@ public class Laser : MonoBehaviour
         }
         if (collider.gameObject.GetComponent<ItemBreak>() != null)
         {
-            collider.gameObject.GetComponent<ItemBreak>().TakeDamage(damage);
+            collider.gameObject.GetComponent<ItemBreak>().TakeDamage((int)damage);
         }
     }
